Added first draft of support for quantity
Original attempt was to have a quantity_left counter which would impose
a temporary limit on the amount users could order while other users had
the widgets in their cart. I am changing back to a simple quantity
counter and allow people to fill their carts and it becomes a first come
first serve but I wanted this in the log so I can come back to it if
necessary.